Styling sort select for large screens
