+++ src/applications/representative-search/components/search/SearchResultsHeader.jsx
@@ -1,13 +1,14 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
+import { sortOptions } from '../../config';
 
 /* eslint-disable camelcase */
 
 export const SearchResultsHeader = props => {
   const { searchResults, pagination, query } = props;
 
-  const { inProgress, context, representativeType } = query;
+  const { inProgress, context, representativeType, sortType } = query;
   const { totalEntries, currentPage, totalPages } = pagination;
 
   const noResultsFound = !searchResults || !searchResults.length;
@@ -45,6 +46,20 @@ export const SearchResultsHeader = props => {
     return 'Results';
   };
 
+  const options = Object.keys(sortOptions).map(option => (
+    <option key={option} value={option}>
+      {sortOptions[option]}
+    </option>
+  ));
+
+  // method for triggering sortResults when sortType updates
+  const handleSortTypeChange = e => {
+    props.updateSearchQuery({
+      id: Date.now(),
+      sortType: e.target.value,
+    });
+  };
+
   return (
     <div className="search-results-header">
       <h2
@@ -70,6 +85,21 @@ export const SearchResultsHeader = props => {
           </>
         )}
       </h2>
+      <div className="sort-dropdown">
+        <label htmlFor="sort-by-dropdown">Sort by</label>
+        <select
+          id="representative-sorting-dropdown"
+          aria-label="Sort"
+          // ref={sortTypeRef}
+          value={sortType}
+          title="Sort by:"
+          onChange={handleSortTypeChange}
+          style={{ fontWeight: 'bold' }}
+        >
+          {' '}
+          {options}{' '}
+        </select>
+      </div>
     </div>
   );
 };
