+++ src/applications/representative-search/sass/find-a-representative.scss
@@ -95,6 +95,40 @@
     #search-results-title {
       margin-bottom: 2em;
     }
+
+    hr {
+      @media (min-width: $large-screen) {
+        width: 64em;
+      }
+    }
+
+    .search-results-header {
+      @media (min-width: $large-screen) {
+        display: flex;
+        align-items: flex-end;
+        // justify-content: center;
+        width: 5000px;
+
+        h2 {
+          margin: 0 7em 0 0 !important;
+          padding: 0px !important;
+        }
+      }
+      .sort-dropdown {
+        label {
+          @media (min-width: $large-screen) {
+            margin-top: 0 !important;
+          }
+        }
+
+        // @media (min-width: $large-screen) {
+        //   position: absolute;
+        //   top: -67px;
+        //   left: 750px;
+        //   width: 271px;
+        // }
+      }
+    }
   }
 
   .primary {
